move `CEditor::OnMouseMove` to `CLayerSelector::UpdateHoveredTiles` Also move related variables and classes.
